properly urlencodes remote songs for audacious
mplayer and firefox both work, but audacious wants urls properly encoded only, and fails without useful feedback when they're not. So Here's a quick wrapper around the python urllib "quote" function, and playrandom modified to use it. Encodes remote songs only (start with "http:", as good an assumption as any)