camera-inserted: small script to move pics off
For me, this was the obvious way to talk gnome into automatically moving all the pictures and such off of an inserted camera card.

see also gnome-volume-manger. And because it was fiendishly difficult to find (no gnome help, and on the second page of google results I finally found it)
%h = hal ID
%m = mount point
%D = device name